Improved build script.
It now tells you what is wrong when it needs to.

+++ biotestmine/setup.sh
@@ -10,6 +10,10 @@ USERPROFILE_DB=$MINENAME-userprofile
 IMDIR=$HOME/.intermine
 PROP_FILE=${MINENAME}.properties
 DATA_DIR=$HOME/${MINENAME}-sample-data
+LOG_DIR=$DIR/log
+LOAD_LOG="${LOG_DIR}/load-data.log"
+PROJECT_BUILD="${DIR}/../bio/scripts/project_build"
+PRIORITIES=$DIR/dbmodel/resources/genomic_priorities.properties
 
 # Inherit SERVER, PORT, PSQL_USER, PSQL_PWD, TOMCAT_USER and TOMCAT_PWD if in env.
 if test -z $SERVER; then
@@ -34,37 +38,50 @@ if test -z $DB_ENCODING; then
     DB_ENCODING=SQL_ASCII
 fi
 
+for dep in perl psql createdb ant; do
+  if test -z $(which $dep); then
+    echo "ERROR: $dep not found - please make sure $dep is installed and configured correctly"
+    exit 1
+  fi
+done
+
+perl -MXML::Parser::PerlSAX \
+     -MText::Glob \
+     -MCwd \
+     -MGetopt::Std \
+     -e 'print "#--- Perl dependencies satisfied\n";'
+
 # Report settings before we do anything.
 if test $DEBUG; then
-    echo SETTINGS:
-    echo " DIR = $DIR"
-    echo " MINENAME = $MINENAME"
-    echo " PROD_DB = $PROD_DB"
-    echo " ITEMS_DB = $ITEMS_DB"
-    echo " USERPROFILE_DB = $USERPROFILE_DB"
-    echo " IMDIR = $IMDIR"
-    echo " PROP_FILE = $PROP_FILE"
-    echo " DATA_DIR = $DATA_DIR"
-    echo " SERVER = $SERVER"
-    echo " PORT = $PORT"
-    echo " PSQL_USER = $PSQL_USER"
-    echo " PSQL_PWD = $PSQL_PWD"
-    echo " TOMCAT_USER = $TOMCAT_USER"
-    echo " TOMCAT_PWD = $TOMCAT_PWD"
-    echo " DB_ENCODING = $DB_ENCODING"
+    echo '# SETTINGS:'
+    echo "#  DIR = $DIR"
+    echo "#  MINENAME = $MINENAME"
+    echo "#  PROD_DB = $PROD_DB"
+    echo "#  ITEMS_DB = $ITEMS_DB"
+    echo "#  USERPROFILE_DB = $USERPROFILE_DB"
+    echo "#  IMDIR = $IMDIR"
+    echo "#  PROP_FILE = $PROP_FILE"
+    echo "#  DATA_DIR = $DATA_DIR"
+    echo "#  SERVER = $SERVER"
+    echo "#  PORT = $PORT"
+    echo "#  PSQL_USER = $PSQL_USER"
+    echo "#  PSQL_PWD = $PSQL_PWD"
+    echo "#  TOMCAT_USER = $TOMCAT_USER"
+    echo "#  TOMCAT_PWD = $TOMCAT_PWD"
+    echo "#  DB_ENCODING = $DB_ENCODING"
 fi
 
-if test ! -d $DIR/log; then
-    mkdir $DIR/log
+if test ! -d $LOG_DIR; then
+    mkdir $LOG_DIR
 fi
 
 if test ! -d $IMDIR; then
-    echo Making .intermine configuration directory.
+    echo '#---> Making .intermine configuration directory.'
     mkdir $IMDIR
 fi
 
 if test ! -f $IMDIR/$PROP_FILE; then
-    echo $PROP_FILE not found. Providing default properties file...
+    echo "#---> $PROP_FILE not found. Providing default properties file..."
     cd $IMDIR
     cp $DIR/../bio/tutorial/malariamine.properties $PROP_FILE
     sed -i "s/PSQL_USER/$PSQL_USER/g" $PROP_FILE
@@ -77,15 +94,15 @@ if test ! -f $IMDIR/$PROP_FILE; then
     sed -i "s/malariamine/$MINENAME/gi" $PROP_FILE
     sed -i "s/localhost/$SERVER/g" $PROP_FILE
     sed -i "s/8080/$PORT/g" $PROP_FILE
-    echo Created $PROP_FILE
+    echo "#--- Created $PROP_FILE"
 fi
 
-echo Checking databases...
+echo '#---> Checking databases...'
 for db in $USERPROFILE_DB $PROD_DB $ITEMS_DB; do
     if psql --list | egrep -q '\s'$db'\s'; then
-        echo $db exists.
+        echo "#--- $db exists."
     else
-        echo Creating $db with encoding $DB_ENCODING ...
+        echo "#---> Creating $db with encoding $DB_ENCODING ..."
         createdb --template template0 \
                  --username $PSQL_USER \
                  --encoding $DB_ENCODING \
@@ -94,54 +111,59 @@ for db in $USERPROFILE_DB $PROD_DB $ITEMS_DB; do
 done
 
 if test -d $HOME/${MINENAME}-sample-data; then
-    echo Sample data already exists.
+    echo '#--- Sample data already exists.'
 else
     cd $HOME
     mkdir $DATA_DIR
     cd $DATA_DIR
     cp $DIR/../bio/tutorial/malaria-data.tar.gz .
-    echo Unpacking sample data...
+    echo '#---> Unpacking sample data...'
     tar -zxvf malaria-data.tar.gz >> $DIR/log/extract.log
     rm malaria-data.tar.gz
 fi
 
 cd $DIR
 if test ! -f project.xml; then
+    echo '#---> Copying over malariamine project.xml'
     cp ../bio/tutorial/project.xml .
-    echo Copied over malariamine project.xml
 fi
-echo Personalising project.xml
+
+echo '#---> Personalising project.xml'
 sed -i "s!DATA_DIR!$DATA_DIR!g" project.xml
 sed -i "s/malariamine/$MINENAME/g" project.xml
 
-PRIORITIES=$DIR/dbmodel/resources/genomic_priorities.properties
 if egrep -q ProteinDomain.shortName $PRIORITIES; then
-    echo Integration key exists.
+    echo '#--- Integration key exists.'
 else
-    echo Adjusting priorities.
+    echo '#---> Adjusting priorities.'
     echo 'ProteinDomain.shortName = interpro, uniprot-malaria' >> $PRIORITIES
 fi
 
 cd $DIR/dbmodel
-echo Building DB
+echo '#---> Building DB'
 ant clean build-db >> $DIR/log/build-db.log
 
-echo 'Loading data (this could take some time) ...'
+if test ! -f $PROJECT_BUILD; then
+    echo "ERROR: Cannot find project build script at $PROJECT_BUILD"
+    exit 1
+fi
+
+echo '#---> Loading data (this could take some time) ...'
 cd $DIR
-../bio/scripts/project_build -b -v $SERVER $HOME/${MINENAME}-dump 2>&1 \
-    | tee -a $DIR/log/load-data.log \
+$PROJECT_BUILD -b -v $SERVER $HOME/${MINENAME}-dump 2>&1 \
+    | tee -a $LOAD_LOG \
     | grep 'action.*took'
 
-echo 'Finished loading data.'
+echo '#--- Finished loading data.'
 cp pbuild.log $DIR/log/
 
 cd $DIR/webapp
-echo 'Building userprofile..'
+echo '#---> Building userprofile..'
 ant build-db-userprofile >> $DIR/log/build-userprofile-db.log
-echo 'Building web-application'
+echo '#---> Building web-application'
 ant default >> $DIR/log/build-webapp.log
-echo 'Releasing web-application'
+echo '#---> Releasing web-application'
 ant remove-webapp release-webapp | tee -a $DIR/log/build-webapp.log | grep tomcat-deploy
 
-echo All done. Logs available in $DIR/log
+echo BUILD COMPLETE: Logs available in $DIR/log
